+++ src/app/LandingPage.tsx
@@ -2,8 +2,8 @@
 
 import Link from "next/link";
 import { Typography } from "@mui/material";
-import { BgWrapper } from "@/components/wrapper/BgWrapper";
-import { FlexWrapper } from "@/components/wrapper/FlexWrapper";
+import { BgWrapper } from "@/src/app/components/wrapper/BgWrapper";
+import { FlexWrapper } from "@/src/app/components/wrapper/FlexWrapper";
 import { landingPageTexts } from "@/constants/landingPage/landingPage";
 
 const styles = {
