+++ src/app/components/rating/Rating.tsx
