+++ src/app/types/product/product.ts
